requirements.txt
web3==6.11.1
eth-account==0.9.0
requests==2.31.0
python-dotenv~=1.0.1